Move devices API into separate module
