+++ crates/native/src/api/media/mod.rs
@@ -1,0 +1,74 @@
+//! Devices and displays managing API.
+
+pub mod constraints;
+
+use std::sync::atomic::{AtomicBool, Ordering};
+
+pub use self::constraints::{
+    AudioConstraints, AudioProcessingConstraints, MediaStreamConstraints,
+    VideoConstraints,
+};
+use crate::{
+    api::{MediaDeviceInfo, MediaDisplayInfo, WEBRTC},
+    devices,
+    frb_generated::StreamSink,
+};
+
+/// Indicator whether application is configured to use fake media devices.
+static FAKE_MEDIA: AtomicBool = AtomicBool::new(false);
+
+/// Configures media acquisition to use fake devices instead of actual camera
+/// and microphone.
+pub fn enable_fake_media() {
+    FAKE_MEDIA.store(true, Ordering::Release);
+}
+
+/// Indicates whether application is configured to use fake media devices.
+pub fn is_fake_media() -> bool {
+    FAKE_MEDIA.load(Ordering::Acquire)
+}
+
+/// Returns a list of all available media input and output devices, such as
+/// microphones, cameras, headsets, and so forth.
+pub fn enumerate_devices() -> anyhow::Result<Vec<MediaDeviceInfo>> {
+    WEBRTC.lock().unwrap().enumerate_devices()
+}
+
+/// Returns a list of all available displays that can be used for screen
+/// capturing.
+#[must_use]
+pub fn enumerate_displays() -> Vec<MediaDisplayInfo> {
+    devices::enumerate_displays()
+}
+
+/// Sets the specified `audio playout` device.
+pub fn set_audio_playout_device(device_id: String) -> anyhow::Result<()> {
+    WEBRTC.lock().unwrap().set_audio_playout_device(device_id)
+}
+
+/// Indicates whether the microphone is available to set volume.
+pub fn microphone_volume_is_available() -> anyhow::Result<bool> {
+    WEBRTC.lock().unwrap().microphone_volume_is_available()
+}
+
+/// Sets the microphone system volume according to the specified `level` in
+/// percents.
+///
+/// Valid values range is `[0; 100]`.
+pub fn set_microphone_volume(level: u8) -> anyhow::Result<()> {
+    WEBRTC.lock().unwrap().set_microphone_volume(level)
+}
+
+/// Returns the current level of the microphone volume in `[0; 100]` range.
+pub fn microphone_volume() -> anyhow::Result<u32> {
+    WEBRTC.lock().unwrap().microphone_volume()
+}
+
+/// Sets the provided `OnDeviceChangeCallback` as the callback to be called
+/// whenever a set of available media devices changes.
+///
+/// Only one callback can be set at a time, so the previous one will be dropped,
+/// if any.
+pub fn set_on_device_changed(cb: StreamSink<()>) {
+    WEBRTC.lock().unwrap().set_on_device_changed(cb);
+}
